\name{dataDesc-package}


\alias{dataDesc-package}
\alias{dataDesc}

\docType{package}

\title{
  dataDesc
}

\description{
  Ensemble de fonctions metier utlisees dans le service MDBS pour la production.
}

\details{
  \tabular{ll}{
    Package: \tab  dataDesc \cr
    Type: \tab
    Package\cr
    Version: \tab  0.1.0 \cr
    Date: \tab  01/10/2022 \cr
    License: \tab neant \cr
    LazyLoad: \tab yes\cr
  }


}

\author{
  Hugo Marthinet  <h.marthinet@gmail.com> \cr
}

\references{

}

\seealso{
  Autres packages developpes au CESPA :

  \code{\link{epiplus}}, \code{\link{lazyMe}}, \code{\link{manageR}},
  \code{\link{surveyscores}}, \code{\link{NSN}}
}

\examples{

}
